move merge-objects into utils
